backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
# AI & NLP
openai>=1.0.0
langchain>=0.1.0
chromadb>=0.4.0
tiktoken>=0.5.0
# Auth
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
# Document Processing
pypdf>=3.15.0
python-docx>=0.8.11
# Utilities
httpx>=0.24.0
pytest>=7.4.0
